move state sync tests into a separate file
