+++ integration-tests/tests/cases/mod.rs
@@ -24,6 +24,7 @@ pub mod mpc;
 pub mod nightly;
 pub mod solana;
 pub mod solana_stream;
+pub mod state_sync;
 pub mod store;
 pub mod sync;
